session_019_DRF_CBV/requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2